moving utils file
