+++ lib/import-single-pattern.js
@@ -9,7 +9,6 @@ var yaml = require('js-yaml'),
 
 var patternCompiler = require('./pattern-compiler.js'),
     cssUtils = require('./css-utils.js'),
-    twigger = require('./../twig-compile').compile,
     utils = require('./../utils');
 
 var File = require('vinyl');
